microbit: Remove the timer for unix flashes, now uFlash always blocks.
Parent commit makes uFlash blocking in all platforms, so we can
remove the 10s timer used in Linux and macOS.
This was required because before the uFlash update it would
return before the micro:bit file transfer was complete.

This also fixes issues that happen when the file transfer takes
longer than 10s and MicroPython is not ready to enter raw REPL
when Mu tries to send main.py via serial.

Fixes #1181